refactor code for better re-use of fns
