+++ gym_fishing/envs/fishing_cts_env.py
@@ -5,10 +5,7 @@ import gym
 from gym import spaces, logger, error, utils
 from gym.utils import seeding
 import numpy as np
-from csv import writer
-from pandas import read_csv, DataFrame
-import matplotlib.pyplot as plt
-
+from gym_fishing.envs.shared_env import harvest_draw, population_draw, csv_entry, simulate_mdp, plot_mdp
 
 class FishingCtsEnv(gym.Env):
     metadata = {'render.modes': ['human']}
@@ -46,21 +43,6 @@ class FishingCtsEnv(gym.Env):
         
         self.action_space = spaces.Box(np.array([0]), np.array([self.K]), dtype = np.float32)
         self.observation_space = spaces.Box(np.array([0]), np.array([2 * self.K]), dtype = np.float32)
-        
-    def harvest_draw(self, quota):
-        ## index (fish.population[0]) to avoid promoting float to array
-        self.harvest = min(self.fish_population[0], quota)
-        self.fish_population = max(self.fish_population - self.harvest, 0.0)
-        return self.harvest
-    
-    def population_draw(self):
-        self.fish_population = max(
-                                self.fish_population + self.r * self.fish_population \
-                                * (1.0 - self.fish_population / self.K) \
-                                + self.fish_population * self.sigma * np.random.normal(0,1),
-                                0.0)
-        return self.fish_population
-
     
     def step(self, action):
       
@@ -68,8 +50,8 @@ class FishingCtsEnv(gym.Env):
         action = np.clip(action, self.action_space.low, self.action_space.high)
         self.harvest = action
         
-        self.harvest_draw(self.harvest)
-        self.population_draw()
+        harvest_draw(self, self.harvest)
+        population_draw(self)
         
         ## should be the instanteous reward, not discounted
         reward = max(self.price * self.harvest, 0.0)
@@ -84,55 +66,26 @@ class FishingCtsEnv(gym.Env):
         
     
     def reset(self):
-        self.fish_population = np.array([self.init_state])
-        self.harvest = self.init_harvest
-        self.action = 0
-        self.years_passed = 0
-        return self.fish_population
-  
+      self.fish_population = np.array([self.init_state])
+      self.harvest = self.init_harvest
+      self.action = 0
+      self.years_passed = 0
+      return self.fish_population
+
+
     def render(self, mode='human'):
-      row_contents = [self.years_passed, 
-                      self.fish_population[0],
-                      self.action,
-                      self.reward]
-      csv_writer = writer(self.write_obj)
-      csv_writer.writerow(row_contents)
-      return row_contents
+      return csv_entry(self)
   
     def close(self):
-      pass
-
+      if(self.write_obj != None):
+        self.write_obj.close()
 
     def simulate(env, model, reps = 1):
-      row = []
-      for rep in range(reps):
-        obs = env.reset()
-        reward = 0
-        for t in range(env.Tmax-1):
-          action, _state = model.predict(obs)
-          row.append([t, obs, action, reward, rep])
-          obs, reward, done, info = env.step(action)
-          if done:
-            break
-        row.append([t+1, obs, None, reward, rep])
-      df = DataFrame(row, columns=['time', 'state', 'action', 'reward', "rep"])
-      return df
-    
-    def plot(self, df, output = "fishing.png"):
-      fig, axs = plt.subplots(3,1)
-      for i in range(np.max(df.rep)):
-        results = df[df.rep == i]
-        episode_reward = np.cumsum(results.reward)                    
-        axs[0].plot(results.time, results.state, color="blue", alpha=0.3)
-        axs[1].plot(results.time.iloc[:-1], results.action.iloc[:-1], color="blue", alpha=0.3)
-        axs[2].plot(results.time, episode_reward, color="blue", alpha=0.3)
+      return simulate_mdp(env, model, reps)
       
-      axs[0].set_ylabel('state')
-      axs[1].set_ylabel('action')
-      axs[2].set_ylabel('reward')
-      fig.tight_layout()
-      plt.savefig(output)
-      plt.close("all")
+    def plot(self, df, output = "results.png"):
+      return plot_mdp(self, df, output)
+      
 
-      
- 
+
+
